refactor($hd-ui): renamed packages to components
