+++ packages/hd-ui/components.json
@@ -1,3 +1,3 @@
 {
-  "hello": "./packages/hello/index.js"
+  "hello": "./components/hello/index.js"
 }
