+++ packages/hd-ui/package.json
@@ -21,9 +21,9 @@
   "style": "lib/theme-chalk/index.css",
   "scripts": {
     "build:file": "node build/bin/iconInit.js & node build/bin/build-entry.js",
-    "build:theme": "node build/bin/gen-cssfile && gulp build --gulpfile packages/theme-chalk/gulpfile.js && cp-cli packages/theme-chalk/lib lib/theme-chalk",
+    "build:theme": "node build/bin/gen-cssfile && gulp build --gulpfile components/theme-chalk/gulpfile.js && cp-cli components/theme-chalk/lib lib/theme-chalk",
     "build:utils": "cross-env BABEL_ENV=utils babel src --out-dir lib --ignore src/index.js",
-    "clean": "rimraf lib && rimraf packages/*/lib && rimraf test/**/coverage",
+    "clean": "rimraf lib && rimraf components/*/lib && rimraf test/**/coverage",
     "dist": "npm run clean && npm run build:file && webpack --config build/webpack.conf.js && webpack --config build/webpack.common.js && webpack --config build/webpack.component.js && npm run build:utils && npm run build:theme",
     "examples:build": "cd examples/hd-ui-example && npm run build && cd .. && cd hd-ui-example-ts && npm run build",
     "examples:install": "cd examples/hd-ui-example && yarn install && cd .. && cd hd-ui-example-ts && yarn install",
