+++ packages/hd-ui/src/index.js
@@ -1,6 +1,6 @@
 /* Automatically generated by './build/bin/build-entry.js' */
 
-import Hello from '../packages/hello/index.js'
+import Hello from '../components/hello/index.js'
 
 const components = [Hello]
